refactor(bug-war-room): modularize war-room state and add smoke tests
