+++ src/views/bug-war-room/components/README.md
@@ -1,0 +1,8 @@
+# Bug War Room Components
+
+Thư mục này được dành riêng để tách các thành phần UI từ `index.vue` trong giai đoạn refactor tiếp theo.
+
+Các thành phần hiện tại bao gồm:
+- `ResultHistoryPanel.vue`
+- `PostmortemPanel.vue`
+- `LearningJournalPanel.vue`
